Add `top_level_only` parameter to `Groups::all` method
This PR adds a `top_level_only` parameter to the `GitLab\Api\Groups::all` method, which limits the results to top level groups, excluding all subgroups.

Because this parameter is not available for the `subgroups` method, it was necessary to split up the `getGroupSearchResolver` method, which was called by both the `all` and the `subgroups` method.