Updatge README.md with additional examples
The only example currently is to create a new project. That's good but I think a lot of people using the library only want to retrieve information about projects, and I found it tricky to figure out how to do that. I've added an example `show()` call to retrieve information about a project, which I found a little tricky, so hopefully including in the README will help someone.